+++ server/grpc/server_test.go
@@ -31,6 +31,7 @@ import (
 	txtypes "github.com/line/lbm-sdk/types/tx"
 	"github.com/line/lbm-sdk/types/tx/signing"
 	authclient "github.com/line/lbm-sdk/x/auth/client"
+	authtypes "github.com/line/lbm-sdk/x/auth/types"
 	banktypes "github.com/line/lbm-sdk/x/bank/types"
 	stakingtypes "github.com/line/lbm-sdk/x/staking/types"
 )
@@ -272,6 +273,75 @@ func (s IntegrationTestSuite) mkTxBuilder() client.TxBuilder {
 	s.Require().NoError(err)
 
 	return txBuilder
+}
+
+func (s *IntegrationTestSuite) TestGRPCCheckStateHeader() {
+	val0 := s.network.Validators[0]
+	authClient := authtypes.NewQueryClient(s.conn)
+	initSeq := uint64(1)
+	AfterCheckStateSeq := uint64(2)
+
+	header := make(metadata.MD)
+	res, err := authClient.Account(
+		context.Background(),
+		&authtypes.QueryAccountRequest{Address: val0.Address.String()},
+		grpc.Header(&header), // Also fetch grpc header
+	)
+	s.Require().NoError(err)
+	var accRes authtypes.AccountI
+	err = s.app.InterfaceRegistry().UnpackAny(res.Account, &accRes)
+	s.Require().NoError(err)
+	s.Require().Equal(
+		initSeq,
+		accRes.GetSequence(),
+	)
+	blockHeight := header.Get(grpctypes.GRPCBlockHeightHeader)
+	s.Require().NotEmpty(blockHeight[0]) // Should contain the block height
+
+	// Broadcast tx to verify gRPC CheckStateHeader
+	txBuilder := s.mkTxBuilder()
+	txBytes, err := val0.ClientCtx.TxConfig.TxEncoder()(txBuilder.GetTx())
+	s.Require().NoError(err)
+	queryClient := txtypes.NewServiceClient(s.conn)
+
+	grpcRes, _ := queryClient.BroadcastTx(
+		context.Background(),
+		&txtypes.BroadcastTxRequest{
+			Mode:    txtypes.BroadcastMode_BROADCAST_MODE_SYNC,
+			TxBytes: txBytes,
+		},
+	)
+	s.Require().Equal(uint32(0), grpcRes.TxResponse.Code)
+
+	// In order for the block to be mined, even a single node requires at least 1~2 seconds, so the sequence number is not yet increased if we query immediately.
+	// So we can validate our CheckState querying logic without `WaitForHeight`
+	ctxWithCheckState := metadata.AppendToOutgoingContext(context.Background(), grpctypes.GRPCCheckStateHeader, "on")
+	res, err = authClient.Account(
+		ctxWithCheckState,
+		&authtypes.QueryAccountRequest{Address: val0.Address.String()},
+	)
+	s.Require().NoError(err)
+	err = s.app.InterfaceRegistry().UnpackAny(res.Account, &accRes)
+	s.Require().NoError(err)
+	s.Require().Equal(
+		AfterCheckStateSeq,
+		accRes.GetSequence(),
+	)
+
+	res, _ = authClient.Account(
+		context.Background(),
+		&authtypes.QueryAccountRequest{Address: val0.Address.String()},
+	)
+	_ = s.app.InterfaceRegistry().UnpackAny(res.Account, &accRes)
+	s.Require().Equal(initSeq, accRes.GetSequence())
+
+	// Wrong header value case. It is deliberately run last to avoid interfering with earlier sequence tests.
+	ctxWithCheckState = metadata.AppendToOutgoingContext(context.Background(), grpctypes.GRPCCheckStateHeader, "wrong")
+	_, err = authClient.Account(
+		ctxWithCheckState,
+		&authtypes.QueryAccountRequest{Address: val0.Address.String()},
+	)
+	s.Require().ErrorContains(err, "invalid checkState header \"x-lbm-checkstate\"")
 }
 
 func TestIntegrationTestSuite(t *testing.T) {